refactor: user와 email 관심사 분리로 controller, service 분리
