+++ src/main/java/com/WearWeather/wear/user/controller/UserController.java
@@ -3,8 +3,6 @@ package com.WearWeather.wear.user.controller;
 import com.WearWeather.wear.global.common.ResponseMessage;
 import com.WearWeather.wear.global.common.dto.ResponseCommonDTO;
 import com.WearWeather.wear.user.dto.RequestRegisterUserDTO;
-import com.WearWeather.wear.user.dto.RequestVerifyCodeDTO;
-import com.WearWeather.wear.user.dto.RequestVerifyEmailDTO;
 import com.WearWeather.wear.user.dto.ResponseDuplicateCheckDTO;
 import com.WearWeather.wear.user.service.UserService;
 import jakarta.validation.Valid;
@@ -13,7 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v1/users")
 @Validated
 @RestController
 @RequiredArgsConstructor
@@ -21,29 +19,18 @@ public class UserController {
 
     private final UserService userService;
 
-    @PostMapping("/users/register")
+    @PostMapping("/register")
     public ResponseEntity<ResponseCommonDTO> signup(@Valid @RequestBody RequestRegisterUserDTO registerUserDTO){
 
         userService.registerUser(registerUserDTO);
         return ResponseEntity.ok(new ResponseCommonDTO(true, "User registered successfully."));
     }
 
-    @GetMapping("/users/nickname-check/{nickname}")
+    @GetMapping("/nickname-check/{nickname}")
     public ResponseEntity<ResponseDuplicateCheckDTO> checkDuplicateNickname(@PathVariable("nickname") String nickname){
 
         userService.checkDuplicatedUserNickName(nickname);
         return ResponseEntity.ok(new ResponseDuplicateCheckDTO(true, ResponseMessage.NICKNAME_AVAILABLE));
     }
 
-    @PostMapping("/email/send-verification")
-    public ResponseEntity<ResponseCommonDTO> verifyEmail(@Validated @RequestBody RequestVerifyEmailDTO verifyEmailDTO){
-        userService.verifyEmail(verifyEmailDTO.getEmail());
-        return ResponseEntity.ok(new ResponseCommonDTO(true, ResponseMessage.SEND_EMAIL));
-    }
-
-    @PostMapping("/email/verify-code")
-    public ResponseEntity<ResponseCommonDTO> verifyEmail(@Validated @RequestBody RequestVerifyCodeDTO verifyCodeDTO){
-        userService.checkEmailAuthCode(verifyCodeDTO.getEmail(), verifyCodeDTO.getCode());
-        return ResponseEntity.ok(new ResponseCommonDTO(true, ResponseMessage.SUCCESS_EMAIL_VERIFICATION));
-    }
 }
