+++ src/main/java/com/WearWeather/wear/user/service/UserService.java
@@ -1,13 +1,11 @@
 package com.WearWeather.wear.user.service;
 
-import com.WearWeather.wear.global.config.RedisConfig;
 import com.WearWeather.wear.global.exception.CustomException;
 import com.WearWeather.wear.global.exception.ErrorCode;
 import com.WearWeather.wear.user.dto.RequestRegisterUserDTO;
 import com.WearWeather.wear.user.entity.User;
 import com.WearWeather.wear.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,9 +17,6 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
-
-    private final MailService mailService;
-    private final RedisConfig redisConfig;
 
     @Transactional
     public void registerUser(RequestRegisterUserDTO requestRegisterUserDTO){
@@ -39,15 +34,9 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public void verifyEmail(String email){
 
-        checkDuplicatedUserEmail(email);
 
-        mailService.sendEmail(email);
-
-    }
-
-    private void checkDuplicatedUserEmail(String email) {
+    public void checkDuplicatedUserEmail(String email) {
 
         Optional<User> existByEmail = userRepository.findByEmail(email);
 
@@ -66,14 +55,5 @@ public class UserService {
         }
     }
 
-    public void checkEmailAuthCode(String email, String authCode){
-        ValueOperations<String, String> valOperations = redisConfig.redisTemplate().opsForValue();
-        String code = valOperations.get(email);
-
-        if(!authCode.equals(code)){
-            throw new CustomException(ErrorCode.NICKNAME_ALREADY_EXIST);
-        }
-
-    }
 
 }
